+++ CHANGELOG.md
@@ -384,6 +384,7 @@ Working external knowledge, personas, preparation for instruments and actions
 -   Allow deep structure (h3, h4,...) in `.ptbk.md` files
 -   Add `SAMPLE` command with notation for parameter samples to `.ptbk.md` files
 -   Add `KNOWLEDGE` command to `.ptbk.md` files
+-   Change `EXECUTE` command to `BLOCK` command
 -   Add `callEmbeddingModel` to `LlmExecutionTools`
 -   `callChatModel` and `callCompletionModel` are not required to be implemented in `LlmExecutionTools` anymore
 -   Remove `MultipleLlmExecutionTools` and make `joinLlmExecutionTools` function